Effects and UI
Pickups on the map now trigger a particle system on the ball. Additionally, the physic material of the ball are modified to simulate buffs when these pickups are obtained.
Furthermore, a progress bar for the scene manager has been implemented. This progress bar provides visual feedback on the current state or progress of the scene management process.